move hideKeyboard out from DynamicTabAdapter
